Readded GameController as a singleton in the game
Stopped the GameController from being created by the GameSetup as it was
failing to find objects in the scene using GameObject.Find() meaning we
couldn't get reference to the menu or spawnpoints from the created
instance.

Currently unable to get a reference between the GameSetup and the
GameController now.
Likely need to use SceneManager.SceneLoaded() to create the reference at
the correct time.