Fixed issue with transfer menu -> game
Settings is now a scriptable object to allow us to set defaults easily.
Need to make it so there are defaults.
Need the menu UI to reflect the defaults of the settings object.
This should fix the current issue where not going into the settings menu
from the gamesetup menu will result in all settings being 0.