Fixed Mouse turning, allowed for infinite time
Mouse movement was looking at the destroy plane which is much lower then
the actual floor and therefore causing issues with depth.
Moving the floor layer to the real floor didn't work because the floor
has holes in it.
Solution was to make an empty game object that encompases the level and
use that as the floormask for the raycast.