Made temp fix for build errors,
Unity won't let you build if a script is 'using UnityEditor' which a
number of scripts are. I have put these in #if blocks, but we will need
to find a different way to get the materials.

Made the GameSetup move back into the menu scene on exit.
Made the GameSetup a singleton.