Fixed Giantdad from spawning.
Stopped the giant players from appearing when a ragdoll was destroyed.
was caused from the destroy plane not destroying the root object, only
the limbs of the ragdoll making it explode, and also removing any
rigidbodies so it stopped moving.

Seperated ragdoll and destory plane into seperate objects, players will
ragdoll when falling out of the level, destroy box is a fair amount
lower.